Add files via upload
This is for the base data gathering process; for exploratory purposes, you can do multiple genres in the genre list, where which can be put into a database locally and into the cloud, however for now, I have not set those environments up for this project so for now, it saves the data as a CSV file, based on the genre and the popularity score. For the analysis for this project but in the future if wanting to evaluate popularity over time.